Fix AC calculation in item description with deformed body (Monsterracer)
Previously, the AC granted by armour when the player has deformed body
was sometimes lower than that indicated in the item description's
"Wearing mundane armour of this type will grant the following:" line.

This was because:
  - player::base_ac_from calculates the AC reduction from
    deformed body as `AC - baseAC / 2`;
  - the description called the player::base_ac_from method with scale
    parameter 1, resulting in the AC being rounded up;
  - the actual AC calculation called the method with scale parameter 100
    and then performed integer division on the final result, resulting
    in the overall AC being rounded down.

Fix this by calling base_ac_from with scale parameter 100 and then
dividing by 100, so that the maths here is equivalent to that in the
real AC calculation.