Adjust an Okawaru altar vault
In the vault okawaru_arena_lexackson, use clear runed doors instead of
opaque ones. Don't use non-standard colors on lava, doors, and down
hatches. Also don't use a special quality level on loot where good_item
would suffice. This vault might predate the availability of that tag,
but the issues described in the now-removed comments are no longer
relevant.

This vault in general should see a rework. It has a spoilery mechanism
to release monsters and block your exit as you move into the vault. It
also has a special "loot choice" mechanism that isn't necessary given
the number of monsters. A transporter vault would be a good fit for the
basic concept, but it would be a little plain if converted to a simple
room with 4 "warrior" monsters. I'll look to rework the vault in a
future commit once I have a good idea.