Adjustments to a runed door vault
In mikee_hates_you, the runed door has a special warning text, but this
isn't necessary. The vault shows what monsters you'll be facing, and the
avatar's mesmerize ability can be seen through x-v. Since the runed
doors are transparent now, move the loot item behind the final door
instead of putting it under the merfolk with no_pickup. Adjust the
windows to allow seeing the monsters as well as the loot item.

The vault places an escape hatch for exit, but allows placement on D:$
where this hatch can't place. Disallow placement on D:$, allowing the
vault to place one level earlier. D:12 is fine for a vault of this kind.
Finally, use the default glyph for runed clear doors instead of a custom
one.