Use a separate rng for ghosts
The most recent set of seed instabilities I've been seeing are because
the ghost loading code makes a different number of random number draws
depending on whether it loads from the permastore or not, and in a clean
setup it was guaranteed to draw from the permastore on the first run,
but not thereafter, for some seeds. (Because even a permastore ghost
gets resaved as an ephemeral ghost if a level is vetoed - something that
is perhaps a bug itself.) In general, it seems risky to use levelgen rng
for something that may depend on what's on a particular file system, so
separate this off into its own rng for cases like this, distinct also
from the gameplay rng.

Also, fixed a bug in loading of ephemeral ghost files where if there was
more than one ghost in the file, the wrong one would be removed after a
ghost was selected for use.