Fix potential iteration order issue for autofight
This is unrelated to my current randomization woes, but something I
noticed while auditing lua/crawl interactions for iter order issues. The
iteration order for lua_next is *not* in general guaranteed, even if the
table is an array, and what I have noticed is that eventually, in a long
execution session, it will start diverging. So this may have been the
source of some of our more obscure reported autofight bugs.

I'm not sure if this solution using `map` is absolutely the best way to
deal with it; lua 5.3 has much better support for doing ipairs-style
iteration in C than lua 5.1, and this was the easiest fix without me
actually figuring out how that's supposed to work in lua 5.1 (which is a
bit tricky from everything I've read).