Update tutorial, fix tutorial/hints mode skill menu text (11851)
Update a traps message that stated traps aren't always revealed, and a
berserk hunger message that stated you can't berserk while very hungry;
this restriction is now limited to near starving or worse.

Shorten the "Read skill descriptions" button description to "Skill
descriptions" to avoid this overlapping the text on the following line.