Explicitly track visited levels
My previous pregen-related commits tried to avoid this, but I don't
think it can be avoided. The basic problem is that this information is
also present in the turncount on a level, but that information is saved
with the level chunk, and it turns out to be kind of useful to have in
many cases where the player isn't on the level in question. Previously,
whether a level was generated (which is tracked in a simple way in the
save package) was synonymous with whether the level had been visited, so
most code just used that. The new implementation uses a CrawlHashTable
as a prop on the player.

This commit specifically fixes three pregen bugs: (i) rune locks were
broken (would always open), (ii) jiyva would eat items on all levels,
not just levels you'd visited, and (iii) one of the perennial problems
with pregen implementations, X] shows a blank screen. It also further
cleans up the morgue vault listing code, and provides more information
for a wizmode morgue. If there are any pregen games in the wild yet,
this fix won't do save compat for them and they'll still have these
bugs, but it will do save compat for regular games.