Add a debugging hook
This is kind of dumb but I can tell I'm going to need this again. This
is aimed at debugging generation that may be happening deep in some test
sequence: toggle this on to echo all dprfs to stderr for some window of
time. (E.g. in a recent episode I used it to print all DNGN_DIAG
messages while the builder was generating Lair:6 for a particular seed
in vault_catalog.lua.)