Merge Piercer into the storm bow
The storm bow is functionally the same as a +8 elec longbow; fairly
strong, but unremarkable. Piercer is a little better since it's a bow
user's only way to get a penetration weapon, but that's still not too
special.

Adding a penetration effect to the storm bow makes it much more
exciting, fits the electric/lightning theme, and puts it almost on par
with the other unrand bow Zephyr. This also makes Piercer obsolete.