Save game seed properly
This was broken in 372e6d594387f, which caused the seed to save as 0
because of the timing of the call to the `player` constructor.  After
initial rng setup this value isn't used for anything meaningful (since
what matters is the rng state itself), so its disappearance was entirely
cosmetic -- but also very confusing to players.