Don't crash when there are game types from the future
The save browser was crashing immediately on starting up in this case,
which is triggered by the presence of seeded games from 0.23. This
future-proofs a bunch of calls used by the save browser so that these
games behave just like any other incompatible future-versioned save game
in the main menu.

(if you want to cherry-pick this to a past version, it is probably
easier to use b7382820b72 from 0.22.)

(cherry picked from commit b7382820b7259ae5d2ffec6fa48112b66c2509ab)