Let player ghosts be shafted (#1003)
This was an artifact from when player ghosts were totally unable to leave the floor they started on by any means. Now that they don't have this trait, and don't broadcast it in their descriptions, there isn't any reason for them to be special-case immune to shafts, or for them to destroy shafts uselessly on moving over them.