Fix cut-off ability cost description (11893)
When the Sacrifice Purity ability is granted to the player with the
inhibited regeneration mutation, the ability's cost description is
  Purity (inhibited regeneration)
which is 31 characters, however the cost string was only allocated 30
characters, so the last parenthesis was being cut off.

Hence, extend the string length for the ability cost description to 32
characters, and adjust the ability screen's title line accordingly.