Don't let overwrite_floor_cell place on transporter markers (11936)
This crash was happening because the abyss entry portal for depths 4
placed on top of a transporter map marker, and those markers get turned
into transporters fairly late in levelgen in a way that unconditionally
overwrites whatever features are there. Then the game crashed on the map
debug scan because a portal is supposed to be guaranteed on this level.

This fix explicitly checks for this kind of marker when placing vaults
that have overwrite_floor_cell; this tag is used for all the fallback
entry vaults for depth, as well as uniq vaults (where this issue doesn't
matter). This solution fixes this exact problem but does point towards
other potential levelgen issues, in that transporters are treated as
floor terrain until the very last minute. I wonder if it might work
better to place transporter features with the markers, and then only do
the connection part at the end of levelgen? But it's also possible that
this won't come up too much, since aside from this case I don't think
there's a way to overwrite any transporter vaults.