Let players disable autopickup from the drop menu
Add a new "drop-and-disable-autopickup" mode for the drop menu, toggled
with '\'. Items selected in this mode are marked with a '*' instead of a
'+' and have their autopickup disabled when the menu is exited.

This turns "drop unwanted item; disable autopickup for unwanted item"
from a two-menu operation into a one-menu operation and makes the
most vital bit of the '\' menu available in a more convenient
location.

InvEntry and SelItem each get a new field to record whether they're
selected "normally" or in this new special way, and InvMenu gets a
field recording which selection mode it's in.  InvMenu now overrides
::pre_process() and ::select_item_index() to handle the new mode.

Also, assorted cleanup/bugfixes:
* Fix '_' on the drop prompt not actually bringing up the help menu.
* _invent_select can't return PROMPT_GOT_SPECIAL, so don't check for it.
* Replace anonymous you.force_autopickup ints with an enum
* Add helper functions for easier access to you.force_autopickup
* Demystify magic -1 to OSEL_ANY in a few places
* Remove unused define PROMPT_INAPPROPRIATE