Linebreak level annotations in chardumps (11904)
This isn't really a complete solution: in general, the chardump doesn't
fully respect 80 chars, this was just the most egregious example. ?# on
tiles is fine with everything else I tested, in that it typically
doesn't expand past the end of the screen, but ?# on console truncates
on the right rather than centering, so still has a bunch of issues. I
think a complete solution needs: (i) on tiles, prevent
formatted_scroller from being wider than the displayable area, (ii)
implement left-right scrolling in formatted_scroller, and (iii) take
another pass to see what can be truncated at 80 characters. (For
some things, like the action table, this probably isn't possible, but
long equipment names for example could be better managed.) I did
consider truncating the entire main chardump at 80 characters but this
tends to look kind of bad in a lot of cases.

This commit also tweaks some wizmode formatting for vault lits so it's
readable in console in ?#.