Don't use div_rand_round for noise values
In wizmode, this could have the following effect: in a non-pregen game,
when a noise happened before descending to a new level, as long as
levelgen triggers a more, drawing the wizmode noise bar could cycle the
levelgen rng while rerendering the hud during that more. (I hope there
aren't more things like this, but there might be, as quite a lot of code
gets called on a more. So it may be worth just forcing the ui rng for
doing mores.) Luckily, non-wizmode noise rendering doesn't call this
code path.