Eliminate some spurious messaging on wizmode level change
If you enter a monster's los, and then exit it off level (e.g. via &d)
instantaneously, and the monster's slot is taken up by a new monster on
the new level, you would be told that this new monster had moved out of
view. This change causes the queue of just seen monsters to be cleared
on a level change, otherwise it required a world_reacts cycle. It's hard
to get this bug to happen without wizmode, but it might have been
possible with serpent's lash and very specific stair placement.

There's still some cases where this messaging interacts badly with
0-turn actions, e.g. serpent's lash on the same level, but there at
least it prints a message about the correct monster.