Add bolt range behaviour to bolt spell descriptions
Add a one-line description of how bolt spells' range is reduced by one
for each monster the bolt strikes, to all relevant bolt and bolt-like
spells and abilities. Until now, this information wasn't documented
anywhere. The description is:

"The range of the [bolt/blast/jet/ball] is decreased by one for every
creature it strikes."

The spells whose descriptions are changed by this commit are:
  Breathe Dispelling Energy
  Breathe Frost
  Breathe Steam
  Breathe Noxious Fumes
  Bolt of Cold
  Bolt of Draining
  Bolt of Fire
  Bolt of Magma
  Chilling Breath
  Cold Breath
  Corrosive Bolt
  Crystal Bolt
  Energy Bolt
  Fire Breath
  Metal Splinters
  Quicksilver Bolt
  Steam Ball
  Venom Bolt.

The bolt-like spells and abilities whose descriptions are not affected
by this commit are:
  Breathe Fire (I'm not actually sure what uses this ability, apart from
                red draconian players, for whom the ability is
		different anyway.)
  Random Bolt
  Shadow Bolt (used for Dithmenos worshippers).

The five lightning-type bolt spells' ranges are not affected in this way
and hence not changed. They are:
  Breathe Lightnng
  Blinkbolt
  Electrical Bolt
  Lightning Bolt
  Shock.