Fix signed casts causing build-specific behavior in random names
Random letter choices in make_name were casting a pseudo-random unsigned
32bit integer to int and then modding it with something of type size_t,
which will usually implicitly cast it back to size_t (which is unsigned
and may vary across platforms; the cast is triggered because it
typically is bigger than the int type so has a higher rank). If the int
value ended up as negative in the first cast, the implicit cast to
size_t was coming out differently on different platforms, leading to
different random names (basically depending on whether it was a 32bit or
64bit build).  This change maintains unsignedness throughout for more
stable behavior.

Because random naming uses its own rng objects, the effect of this was
partitioned off from anything else in levelgen, leading to identical
properties on randarts with different names across platforms in gotm 2
(since windows is a 32bit build).