Update the bounded random implementation
This moves to an implementation of unbiased bounded random number
generation based on a recent algorithm by Daniel Lemire with an
implementation by Melissa O'Neill. This change has two advantages: (i)
the big one, this is off-the-shelf code written by an rng expert,
implementing an algorithm designed by someone who specializes in this
kind of optimization. Ours was *probably* fine, but didn't exactly match
any standard way of doing this that I was able to find.  (ii) It gets a
10% speedup for typical range values compared to our old implementation,
because it avoids using /,% where possible. This is probably not a big
deal in the scale of things that are slow in crawlcode, but we do call
the rng a *lot*, mostly with range values that this function is
optimized for. See:

    http://www.pcg-random.org/posts/bounded-rands.html
    https://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
    https://arxiv.org/abs/1805.10941

This also updates the API for this call, moving it the implementation
into the pcg class, as well as updates some comments in pcg.cc to be
more detailed and correct.