Prevent deep water teleport closets in swamp (9880)
This was recently reported here:
   https://www.reddit.com/r/dcss/comments/bn20bv/ive_never_seen_this_before_is_it_normal_to_have/

and thanks to seeding I was able to reliably replicate this bug from the
morgue posted in that thread. I'll upload a save file to 9880 for
posterity. (I'm assuming that 9880/11519 is the same bug and that
players will tend not to notice/report these cases unless there's a
monster in the closet.)

Basically, what happens is this: there were two connectivity passes for
swamp. The first was via zonify, and it finds and successfully fills in
any teleport closets existing at the time with trees -- but does not
check for overall level connectivity involving deep water. Then there's
a c++ pass in dungeon.cc:_build_vault_impl that is swamp-specific. This
pass handles connectivity issues involving deep water. Unfortunately,
because it treats deep water as impassable, it does not fill in deep
water squares, creating teleport closets of the kind that the zonify
pass fixes. (Also, I've verified that these are teleport closets for
flying/swimming chars.)

To fix this I have inelegantly added a *third* pass on the c++ side,
that finds and fills in any resulting deep water tele closets. It might
be possible to do without the zonify pass now, since this (now) third
pass would handle these cases, but I'm worried about kicking the tower
of cards. Ideally, this would all be handled by zonify, but right now I
don't understand zonify well enough to get this to happen.