Clear mon_acting on _reset_game (11971)
If the player is killed at the same time as other monsters under certain
conditions (I replicated it with inner flame), mon_acting would not get
cleaned up. This could lead to a crash when starting a new game without
exiting the program, as long as a monster would spawn in los.