Don't crash on gozag + gold in grunt_nemelex_the_gamble
Unobtainable items don't appear in the stash tracker, and calls that try
to add them to it silently fail.  The only remaining "unobtainable"
items are in grunt_nemelex_the_gamble; these are marked unobtainable so
that if the player rolls the peril room they don't show up in the stash
tracker. However, Gozag's gold detection code didn't take this flag into
account, triggering an ASSERT on the expectation that any gold generated
at all would be trackable.

This isn't the most elegant solution, but it fixes the crash. I think
what would be better is for (i) "unobtainable" to be removed, since it
appears to have been deprecated for everything except this vault some
time ago, and (ii) for this vault to do something like destroy the items
if the player lands in the peril room.