Disable pattern matching against levelgen output
On a debug build, this was the heaviest single code path after
luaV_execute itself, and this output consists entirely of debug output
that should be hidden anyways. If someone had a use case for leaving
these enabled I can try to come up with other solutions, but right now
I'm not seeing anything crucial.