Don't use locale-sensitive string comparison for table.sort
This is leading to seed divergence on different systems depending on
locale, most immediately because some locales collate capital letters
with their lowercase versions, and some don't. Lua uses locale-sensitive
string comparisons unless you override it at build time, and it's not
practical to set the locale in a way that avoids this (or at least I
haven't managed). Capitalization isn't the only issue that could come up
here, either, and it isn't appropriate for internal data to be affected
by locale in this way (wtf lua), so this commit just exports c++ string
comparison and uses that instead, which should use lexicographic byte
ordering.