Allow running clua from map viewing mode (#1103)
This is made possible by adding user function recognition to key inputs while viewing map. A clua function can be run by adding a levelmap key map from a given key to ===functionName. When the clua function finishes, we will still be in map mode unless the function took us out of it.

Additionally add clua functions for asking about map mode: crawl.in_map_mode, crawl.cursor_pos, and crawl.cursor_level

cursor_pos and cursor_level return nil outside of map mode. cursor_pos also returns nil if the level of the cursor is not the same as the level of the player, because giving the relative position to the player in that case would leak information about the level map.