Allow using G_ to select altar for travel
Pressing _ in G now lists discovered altars, user can then select the
altar for travel by entering the initial of a god (except for TSO, which
is replaced by 1). Once a god is selected, travelling to the nearest
altar is started.

Help file is also modified to reflect this change.