Add _god_from_initial() to travel.cc
_god_from_initial(const char god_initial) returns a god_type enum
when the initial of a god (i.e., god_initial, case-insensitive) is given.
(except for TSO, which is replaced by '1'). If the given letter does not
match any god's initial, GOD_NO_GOD will be returned instead.

For example: _god_from_initial('a') or _god_from_initial('A')
will return GOD_ASHENZARI.

This function is later used for allowing players to select an altar
for travel by the initial of a god.