Fix a potential cursor rendering issue in glyph mode
What was happening here was if w or h were non-integer, then the scale
value ended up slightly different than the cell sizes. In glyph mode,
this was happening to the width parameter depending on font details
(applying at least to firefox/mac on default settings). When rendering
the mouse cursor, this scale value is used, so it would be a bit larger
horizontally than a single cell, by 1-2 pixels. On firefox this meant
that the cursor didn't get cleanly cleared when it moved, leaving
vertical lines when moving the mouse cursor across the dungeon view.

In testing, I have seen no consequences of using a scale based on the
floored versions of w and h, even though in principle this might mean
that the width of glyphs is slightly larger than the cell size is
expecting.