Don't crash if redrawing a prompt triggers a flash
This is probably a webtiles-specific bug: if a prompt triggered a flash
(on just webtiles builds, I think), and there were monsters in view,
then the redraw involved in the flash would lose the cursor position and
leave the region in the mlist.

This is a kind of minimalistic fix in that it only addresses flash calls
specifically -- viewwindow on relevant builds (as well as
update_monster_pane) can change the region as a side effect, leaving it
in the monster list. However, I'm not yet certain whether it would have
other consequences to adjust this more generally. But probably a more
general fix is needed, and all the occasional cursor out of bounds
issues we see with GOTO_MLIST do result from this side effect of
viewwindow.