Default disable incremental pregen for testing
This is for the merge of the incremental_pregen branch: the branch
includes a wide range of rng improvements etc that are independent of
the generation method, so it will be simpler to test the branch online
without incremental pregen being enabled by default. If players are
interested, they can still enable it in their rc file
(`incremental_pregen=true`). Later, I'll default this to true.