Remove probably-no-longer-needed kludge
The handle_terminal_resize() call in startup.cc (added in e03bc23d4eded)
seems to no longer be required, since the code to determine the HUD size
based on species (added in the same commit) is no longer present, and
thus the HUD should not need to be resized any more. This permits some
minor simplification of handle_terminal_resize().