Allow querying current scissor without UI layout
Widgets may query the current scissor region to reduce render calls,
e.g. by not rendering children or other elements that are not visible.
During a regular layout and render cycle, the UI framework pushes a
scissor region of the entire screen, so this works fine. When doing
ad-hoc rendering (while we transition to full use of the UI framework),
this is not the case, which caused crashes.