Fix monster list position when re-running arena
The call to expand_mlist() assumes that the monster list has its initial
position and size; this is not the case if you run arena mode more than
once without quitting the game in between.

After this fix, the UI is still broken if you run arena then a regular
game. I tried using an unwind_var, but that caused crashes, and a better
fix is in the works anyway (using the widget toolkit).