Simplify macro_buf_add_with_keymap
We only ever add a single key at a time to the macro buffer; better to
enforce that.

This also asserts in remap_key() that the macro buffer is empty before
remapping a key; this must be the case because ui::pump_events() first
tries to get a key from the macro buf, and only remaps a new key if
there is no macro key available.