Slightly improve console MenuButton highlighting code
This now handles another bug by generalizing slightly from 5a66b1aa3b36:
the weapon selection menu wasn't highlighting correctly in local console
because those buttons on console consist of a bin with two Texts. The
new version finds the first Text child (if any) and does the
highlighting on it, and because of this also does away with the #ifdefs.