Fix newgame menu highlight width on remote console
On TILEWEB=y console games, the text widget was not expanding to take
the full width of the box it was contained in. When its background
colour was changed, only the section of the menu button with letters
changed colour.

This isn't a problem in pure console builds because in that case the
text widget is the direct child of the MenuButton widget and so takes
its full size.