Fix difficult-to-read main menu labels on console (advil)
Menu labels previously relied on libunix.cc's adjustment of identical
foreground and background colours, which didn't work properly with
allow_extended_colours = true, since white and lightgrey are not
identical.

This commit manually adjusts all button label colours on focus change,
both foreground and background. This fixes a bug on the weapon selection
menu on console where the background did not extend under the listed
aptitudes, due to them being in a separate text widget. It also cleans
up the hacky code that was previously used to recolour the buttons.

You can specify the foreground colour to use when focused; the default
is black. This code assumes that all text widgets have a single colour,
and will not correctly restore multi-coloured text on unfocus.