Use quads for line drawing (advil)
Line drawing with GL_LINES seems to be subject to a number of rendering
bugs and limitations that vary by OS and driver. This commit renders all
lines with a single quad. This removes support for arbitrary lines that
are not horizontal or vertical, but we don't currently draw any such
lines.