De-randomize Golden Eye confusion spellpower
This random power is a legacy dating back to their creation. The commit
log doesn't describe why Confusion Gaze should work this way instead of
with a normal confuse spellpower and a normal MR check. Now that
monster spell success chance is displayed, this randomization had the
side effect of making the success chance display wrong.

Confusion Gaze was the last monster spell with monster spellpower
randomization; this commit cleans up some of the code, but there is a
larger refactor possible that is left undone.

Closes #1160