Remove ui::Widget align_self property
This only did anything inside a box widget, and was confusingly named
(does it refer to main- or cross-axis alignment?). It seems to be
unnecessary, although relying on the parent stretch property is probably
slightly brittle.

IIRC Flutter allows the parent to store a small amount of data on child
widgets; perhaps adding a Box::set_child_cross_alignment() that utilizes
this method will work well.