Don't scale when rendering lines (Goratrix, 12085)
This bug was caused by line buffer rendering correctness depending on
the GL scale being the same when buffer packing and when rendering,
which was only required due to the way the local tiles minimap worked.

The local tiles minimap has been changed to render the boundary box with
no scaling, fixing this weird quirk.